+++ examples/20_calculate_farm_power_with_uncertainty.py
@@ -30,24 +30,32 @@ historical SCADA data and robust optimization.
 """
 
 # Instantiate FLORIS using either the GCH or CC model
-fi = FlorisInterface("inputs/gch.yaml") # GCH model
-fi_unc = UncertaintyInterface("inputs/gch.yaml") # Add uncertainty with default settings
+fi = FlorisInterface("inputs/gch.yaml")  # GCH model
+fi_unc = UncertaintyInterface("inputs/gch.yaml")  # Add uncertainty with default settings
 
 # Define a two turbine farm
 D = 126.0
-layout_x = np.array([0, D*6, D*12])
+layout_x = np.array([0, D * 6, D * 12])
 layout_y = [0, 0, 0]
 wd_array = np.arange(0.0, 360.0, 1.0)
-fi.reinitialize(layout_x=layout_x, layout_y=layout_y, wind_directions=wd_array)
-fi_unc.reinitialize(layout_x=layout_x, layout_y=layout_y, wind_directions=wd_array)
+wind_speeds = 8.0 * np.ones_like(wd_array)
+fi.reinitialize(
+    layout_x=layout_x, layout_y=layout_y, wind_directions=wd_array, wind_speeds=wind_speeds
+)
+fi_unc.reinitialize(
+    layout_x=layout_x, layout_y=layout_y, wind_directions=wd_array, wind_speeds=wind_speeds
+)
+
+# Set the uncertain
+fi_unc.set_uncertain()
 
 # Define a matrix of yaw angles to be all 0
 # Note that yaw angles is now specified as a matrix whose dimesions are
 # wd/ws/turbine
-num_wd = len(wd_array) # Number of wind directions
-num_ws = 1 # Number of wind speeds
-num_turbine = len(layout_x) #  Number of turbines
-yaw_angles = np.zeros((num_wd, num_ws, num_turbine))
+num_wd = len(wd_array)  # Number of wind directions
+num_ws = 1  # Number of wind speeds
+num_turbine = len(layout_x)  #  Number of turbines
+yaw_angles = np.zeros((num_wd, num_turbine))
 
 # Calculate the nominal wake solution
 fi.calculate_wake(yaw_angles=yaw_angles)
@@ -61,10 +69,10 @@ farm_powers_unc = fi_unc.get_farm_power() / 1e3
 
 # Plot results
 fig, ax = plt.subplots()
-ax.plot(wd_array, farm_powers_nom.flatten(), color='k',label='Nominal farm power')
-ax.plot(wd_array, farm_powers_unc.flatten(), color='r',label='Farm power with uncertainty')
+ax.plot(wd_array, farm_powers_nom.flatten(), color="k", label="Nominal farm power")
+ax.plot(wd_array, farm_powers_unc.flatten(), color="r", label="Farm power with uncertainty")
 ax.grid(True)
 ax.legend()
-ax.set_xlabel('Wind Direction (deg)')
-ax.set_ylabel('Power (kW)')
-plt.show()
+ax.set_xlabel("Wind Direction (deg)")
+ax.set_ylabel("Power (kW)")
+# plt.show()
